acronym-completion-platform/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
google-generativeai==0.3.1
requests==2.31.0
psutil==5.9.5
bcrypt==4.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
flake8==6.1.0
black==23.9.1

# Testing dependencies
httpx==0.25.1
aioresponses==0.7.4 